Address PR feedback: remove dependency on github.com/prometheus/prometheus.
